Split the FavoriteToggleController in 2 reusable controllers
Since at the moment we have two actions that respond to the Show/Hide button click, let's split the FavoriteToggleController in a CssController and a TextController, one responsible of toggling the `hidden` class to the favorite concerts div, and the other responsible for changing the Show/Hide text on the button.

For the CssController we pass the CSS class from the view instead of hardcoding it on the controller, thanks to the data class `data-css-css-class`. We have a boolean statusValue to toggle the class when is true.

For the TextController we also extracted the hardcoded values for the Show/Hide strings to the View, and we have a statusValue also here, so that we can change the button text depending on the statusValue.